Converted my Rendering to an Arbitrary Loop Function
Can't believe that worked! I love these types of functions!
Considering the fact that I don't really need one of these since I have a fixed amount of items to render, I figured considering the nature of the project that I will need to come back and add more parameters. It's kind of a meta-programming move, a function made specifically to make programming easier, the ultimate future proof!
That was over-dramaticized, but the hype is justified, I love arbitrary parametrs in functions!!